+++ autonomy/cli/helpers/chain.py
@@ -32,7 +32,12 @@ from aea.helpers.base import IPFSHash
 from texttable import Texttable
 
 from autonomy.chain.base import ServiceState, UnitType
-from autonomy.chain.config import ChainConfigs, ChainType, ContractConfigs
+from autonomy.chain.config import (
+    ChainConfigs,
+    ChainType,
+    ContractConfig,
+    ContractConfigs,
+)
 from autonomy.chain.constants import (
     AGENT_REGISTRY_CONTRACT,
     COMPONENT_REGISTRY_CONTRACT,
@@ -185,6 +190,25 @@ class OnChainHelper:  # pylint: disable=too-few-public-methods
 
         return ledger_api, crypto
 
+    def check_required_enviroment_variables(
+        self, configs: Tuple[ContractConfig, ...]
+    ) -> None:
+        """Check for required enviroment variables when working with the custom chain."""
+        if self.chain_type != ChainType.CUSTOM:
+            return
+        missing = []
+        for config in configs:
+            if config.contracts[self.chain_type] is None:
+                missing.append(config)
+
+        if len(missing) == 0:
+            return
+
+        error = "Addresses for following contracts are None, please set them using their respective environment variables\n"
+        for config in missing:
+            error += f"- Set `{config.name}` address using `CUSTOM_{config.name.upper()}_ADDRESS`\n"
+        raise click.ClickException(error[:-1])
+
 
 class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
     """Mint helper."""
@@ -243,18 +267,27 @@ class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
             contract_address = ContractConfigs.get(
                 SERVICE_REGISTRY_CONTRACT.name
             ).contracts[self.chain_type]
+            self.check_required_enviroment_variables(
+                configs=(ContractConfigs.service_registry,)
+            )
         elif self.package_type == PackageType.AGENT:
             is_service = False
             is_agent = True
             contract_address = ContractConfigs.get(
                 AGENT_REGISTRY_CONTRACT.name
             ).contracts[self.chain_type]
+            self.check_required_enviroment_variables(
+                configs=(ContractConfigs.agent_registry,)
+            )
         else:
             is_service = False
             is_agent = False
             contract_address = ContractConfigs.get(
                 COMPONENT_REGISTRY_CONTRACT.name
             ).contracts[self.chain_type]
+            self.check_required_enviroment_variables(
+                configs=(ContractConfigs.component_registry,)
+            )
 
         self.old_metadata = resolve_component_id(
             ledger_api=self.ledger_api,
@@ -357,6 +390,18 @@ class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
         component_type: UnitType = UnitType.COMPONENT,
     ) -> None:
         """Mint component."""
+
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.registries_manager,
+                (
+                    ContractConfigs.component_registry
+                    if component_type == UnitType.COMPONENT
+                    else ContractConfigs.agent_registry
+                ),
+            )
+        )
+
         try:
             self.token_id = _mint_component(
                 ledger_api=self.ledger_api,
@@ -405,6 +450,18 @@ class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
     ) -> None:
         """Mint service"""
 
+        if self.chain_type == ChainType.CUSTOM and token is not None:
+            raise click.ClickException(
+                "Cannot use custom token for bonding on L2 chains"
+            )
+
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.service_manager,
+                ContractConfigs.service_registry,
+            )
+        )
+
         try:
             token_id = _mint_service(
                 ledger_api=self.ledger_api,
@@ -442,6 +499,16 @@ class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
 
     def update_component(self, component_type: UnitType = UnitType.COMPONENT) -> None:
         """Update component."""
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.registries_manager,
+                (
+                    ContractConfigs.component_registry
+                    if component_type == UnitType.COMPONENT
+                    else ContractConfigs.agent_registry
+                ),
+            )
+        )
         try:
             self.token_id = _update_component(
                 ledger_api=self.ledger_api,
@@ -455,7 +522,7 @@ class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
             raise click.ClickException(f"Invalid parameters provided; {e}") from e
         except ComponentMintFailed as e:
             raise click.ClickException(
-                f"Component mint failed with following error; {e}"
+                f"Component update failed with following error; {e}"
             ) from e
 
         click.echo("Component hash updated:")
@@ -481,6 +548,24 @@ class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
     ) -> None:
         """Update service"""
 
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.service_manager,
+                ContractConfigs.service_registry,
+            )
+        )
+
+        *_, state, _ = get_service_info(
+            ledger_api=self.ledger_api,
+            chain_type=self.chain_type,
+            token_id=cast(int, self.update_token),
+        )
+
+        if ServiceState(state) != ServiceState.PRE_REGISTRATION:
+            raise click.ClickException(
+                "Cannot update service hash, service needs to be in the pre-registration state"
+            )
+
         try:
             token_id = _update_service(
                 ledger_api=self.ledger_api,
@@ -501,7 +586,7 @@ class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
             )
         except ComponentMintFailed as e:
             raise click.ClickException(
-                f"Service mint failed with following error; {e}"
+                f"Service update failed with following error; {e}"
             ) from e
 
         click.echo("Service updated with:")
@@ -539,6 +624,11 @@ class ServiceHelper(OnChainHelper):
         token: Optional[str] = None,
     ) -> "ServiceHelper":
         """Check if service"""
+        if self.chain_type == ChainType.CUSTOM:
+            self.token = token
+            self.token_secured = False
+            return self
+
         self.token = token
         self.token_secured = is_service_token_secured(
             ledger_api=self.ledger_api,
@@ -584,6 +674,13 @@ class ServiceHelper(OnChainHelper):
                 spender=spender,
             )
 
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.service_manager,
+                ContractConfigs.service_registry,
+            )
+        )
+
         try:
             _activate_service(
                 ledger_api=self.ledger_api,
@@ -619,6 +716,13 @@ class ServiceHelper(OnChainHelper):
                 spender=spender,
             )
 
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.service_manager,
+                ContractConfigs.service_registry,
+            )
+        )
+
         try:
             _register_instance(
                 ledger_api=self.ledger_api,
@@ -642,6 +746,16 @@ class ServiceHelper(OnChainHelper):
     ) -> None:
         """Deploy a service with registration activated"""
 
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.service_manager,
+                ContractConfigs.service_registry,
+                ContractConfigs.gnosis_safe_proxy_factory,
+                ContractConfigs.gnosis_safe_same_address_multisig,
+                ContractConfigs.multisend,
+            )
+        )
+
         try:
             _deploy_service(
                 ledger_api=self.ledger_api,
@@ -660,6 +774,13 @@ class ServiceHelper(OnChainHelper):
     def terminate_service(self) -> None:
         """Terminate a service"""
 
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.service_manager,
+                ContractConfigs.service_registry,
+            )
+        )
+
         try:
             _terminate_service(
                 ledger_api=self.ledger_api,
@@ -674,6 +795,13 @@ class ServiceHelper(OnChainHelper):
 
     def unbond_service(self) -> None:
         """Unbond a service"""
+
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.service_manager,
+                ContractConfigs.service_registry,
+            )
+        )
 
         try:
             _unbond_service(